data_pipeline/pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "openai>=2.13.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
